dma-heap: qcom: Introduce "qcom,unmapped" flag for pre assigned regions
The secure carveout heap driver assumes that a region of memory can
be freed and assigned to a secure domain by the OS. However, in
scenarios where the OS does not have the sufficient privileges to
meddle with the memory, the OS can rely on regions of memory that
have been pre-assigned to a secure domain.

Introduce a flag "qcom,unmapped" to indicate that a region of memory
has been pre assigned to a secure domain. Use this flag to prevent
the secure carveout heap from zeroing out or assigning this memory.

Change-Id: Ic52250a6f9d9845430a20906b2f84ab6e8bbf176
Signed-off-by: Mahesh Kumar <quic_mahkumar@quicinc.com>
(cherry picked from commit 6f9e69835adc84d1f42ebce9c14fc7c19451560d)