sched/walt: Skip Walt Vendor Hook Handling for SCX Tasks
Scx tasks will be handled by the designated scx scheduler. WALT will
not handle scx tasks, so skip walt and handle scx tasks through the
vendor hook.

In some walt vendor hooks, the judgement of whether a task is an
scx task is placed before the check for walt_disabled. This is
because, in future scenarios, the likelihood of hitting the
condition to judge if a task is an scx task is higher, thereby
reducing the execution of unnecessary instructions.

Change-Id: I4fddfcd33c6e9d4fd19fdc78747e1ca5c84e928d
Signed-off-by: Tengfei Fan <quic_tengfan@quicinc.com>