sched/walt: Skip walt ops and busy time updates when SCX is enabled
Skip WALT enqueue, dequeue, and setting task CPU if the task is an
SCX task when SCX is enabled.
Skip updating the WALT busy time of SCX tasks when they need to be
added to or removed from the WALT task group. Since SCX tasks may
still switch to a WALT-managed scheduling class, adding or removing
SCX tasks from the WALT group is still allowed when SCX is enabled.
However, SCX tasks should not contribute to WALT busy time statistics.

Change-Id: I3585458ce4368f351cb695a60f4998b1e13aef46
Signed-off-by: Tengfei Fan <quic_tengfan@quicinc.com>