drivers: dcvs: bwmon: Add max_mbps check for max_freq
Add a new tunable parameter called max_freq_max_mbps
which is used to specify the maximum mbps for which
max_freq should be used. More specifically, when the
traffic seen exceeds this new threshold, the max_freq
limitation will be bypassed and hw_max_freq will be
used instead. This allows the IB vote to not be limited
in cases of very high traffic.

Change-Id: I7abc3cc2e867afac5aed62a4748e137d030e44c4
Signed-off-by: Amir Vajid <quic_avajid@quicinc.com>