minidump: Fix logic for slab owner collection
Previous logic allowed for a case where SLAB_STORE_USER was not a part
of the kmem_cache flags. This caused an issue when trying to retrieve
the alloc stacks for the slabs because there was no tracking data
stored. Change this to make sure that the flag is set on the kmem_cache
before attempting to get the tracking information.

Also change CONFIG_SLUB_DEBUG_ON to CONFIG_SLUB_DEBUG as this seems to
be the more correct config to check in this case.

Change-Id: Idb51dcd4f7954fc3ef77e6a2c7558a220cccbdc9
Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>