net: pcs: Prevent unintended error handling in USXGMII mode
In the PCS driver, the switch-case for PHY interface modes
previously allowed fallthrough from 10GBASER and 5GBASER to
USXGMII, which could lead to unintended execution of the
error handling.

To address this, replaced the `break` statement with a `return`
in the USXGMII case,which ensures the control exits the function
immediately after handling USXGMII mode, preventing accidental
execution of error handling logic meant for earlier cases.

Change-Id: Id912109086bc00df7af0f95bb104c4bc6cadf603
Signed-off-by: Uppalamarthi Sowmya <quic_usowmya@quicinc.com>