sched/walt: fix small task on large cpus for obet
Currently, obet only relies on a big task getting enqueued
to register its pid i.e. an enqueue of small task is ignored.

However if a small task is enqueued, and it runs on the large cpu
long enough to become a big task, obet fails to register this pid.

To fix this, create a flag to force reeval of big pid at the window
rollover checkpoint i.e. check_obet(). Set this flag everytime
the util of a task that is running on a large cpu increases.

Note that the update_history() of a task will be called prior
to check_obet() in the window rollover handling sequence.
Hence the setting of the flag in update_history() will
be properly consumed in the check_obet().

Change-Id: I57d6317d1179060502bab86ad9d4c59a9b51adc6
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>