sched/walt: Correct LST tracker type and enhance pipeline tracing
The LST tracker 'lst_state_counter', which determines the delay for task
pipeline selection was previously defined as an unsigned integer. This
causes an unintended overflow if the counter is decremented past zero
when a task becomes active, resulting in big positive values due to
wraparound. To address this, change the type to a signed integer.

while at it, remove continuous activity based 'lst_state_counter'
manipulation, the counter is reset once we complete the LST period
after that the whole calculation restarts, this simplifies code.

Additionally, enhance the 'sched_task_util' tracepoint to provide more
pipeline related information. This added visibility will aid in
debugging pipeline issues.

Change-Id: I385d3c63687eb7df4fbd21650a78215f09113bf6
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>