net: stmmac: Change PDMA mapping to resolve SMMU crash
Change mapping from PDMA7 <-> TC5 to PDMA7 <-> TC6 to match SMMU
to avoid SMMU crash.

Previously we mapped queue 5/6/7 to TC5, but queue 7 is not enabled.
so we mapped only queue 5/6 to TC5, mapped queue7 to TC6.
once the packet goes to TC5, it won't go to queue 7 and cause crash.

Change-Id: I0c9ab6cedb9acf1accef4fa9f66e98755779b4c1
Signed-off-by: Mei Xie <quic_meixie@quicinc.com>
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>