sched/walt: Ensure that SBT and TB are off when fmax is reduced
In situations where the large cluster's fmax is manually reduced, small
tasks can upmigrate to prime and lead to triggering of trailblazer and
SBT conditions unnecessarily.
To prevent this, add a condition in the SBT and trailblazer state update
paths which checks whether the Large cluster has reduced capacity beyond
a particular threshold. And if so, prevent the system from entering SBT
or trailblazer state.

Change-Id: If61be4aa1ede8b7ff72db65d01ca1b602d3e1644
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>