sched/walt: Add hysteresis to oscillate condition for bus dcvs boost
Consider the situation where bwmon checks whether bus dcvs needs to be
boosted and sees that oscillation is not active and hence does not
boost. Next, oscillation is triggered via FAST interrupt, but before the
next bwmon check occurs, FAST also sends an interrupt to stop
oscillation. So, even though the task has oscillated once between the
Large CPUs, bwmon does not boost bus dcvs.
To reduce the amount of time bus dcvs is not boosted during oscillation,
due to occurrences of the above, introduce a hysteresis of 16ms to the
bus dcvs oscillation boost signal.

Change-Id: I9c8b2d3c2360bcaa92d23fb7a7d1470a338c0952
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>