sched/walt: Allow dynamic update of number of zones
Previously, the number of zones was treated as write-once, preventing
updates after initial configuration. This change removes that
restriction, enabling the number of zones to be modified at runtime.

Note: If the user updates the number of zones and then sets the
corresponding target load values, any previously configured values for
the old zones will be reset.

Change-Id: Iffbeeb41d4b2144f18e59b16691e212aa2d8c6f5
Signed-off-by: Ankit Sharma <quic_anshar@quicinc.com>