power: supply: qti_battery_charger: Update probe() bail out sequence
The PMIC Glink client is used in the work thread functions as well as
the notification callbacks, it should be unregistered at the end of the
bail out sequence after all the work threads have been cancelled and
notifiers unregistered. Update the bail out sequence in probe()
function accordingly, and make the same update in remove() function.

Change-Id: I258ee82ac8d04e37642fa390c784a73d1d34d734
Signed-off-by: Fenglin Wu <fenglin.wu@oss.qualcomm.com>