sched/walt: Keep track of giant tasks enqueued in last window
Track the number of giant tasks enqueued in the last window for at least
a significant portion of it by calculating a product-sum of number
of tasks enqueued and for how long they were enqueued.
This number will be used in a subsequent patch to determine if Big Task
Rotation should be enabled or not.
Note that product-sum is not updated at the time of marking/unmarking
of tasks as giant in the UTRA path as it is going to be updated in the
upcoming window rollover.
The giant task classification and its statistics will be used in
following patches to revisit Big Task Rotation determination.

Change-Id: Ic4a76ea393b2b4ef4cf1534d925e50206c9f8edd
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>