sched/walt: Update BTR determination
Determination of BTR utilizes upmigration percentage as a threshold to
decide if a task is considered big or not. This leads to the issue of a low
upmigration percentage triggering BTR unnecessarily.
Remove the dependency of BTR determination on the user-centric upmigration
percentage by using the giant task classification to determine if a task
is big.

Change-Id: I5cb9f71fda8d8e956bc2bd0ff197fc0590de3853
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>