sched/walt: Introduce giant tasks and their counters
Determination of Big Task Rotation and high performance frequency cap
features utilizes upmigration percentage as a threshold to decide if a task
is considered big or not. This leads to the issue of a low upmigration
percentage classifying tasks as big and subsequently triggering these
features.
Remove the dependency on the user-centric upmigration percentage by using
the giant task classification method to determine if a task is big.
A giant task is one whose util exceeds a particular fraction of the mid
cluster CPUs' capacity_orig.
At every demand update for a task, check if it qualifies as a giant task.
Accordingly, mark or unmark the task as a giant and update per rq counters
which keep track of the number of such giant tasks enqueued on them.
Note that a task's giant status may be stale up until the point where an
update_history runs for the task and has the chance to make its giant
status current again. For eg., in a situation where a small task marked as
giant in a thermal scenario (due to suppression of capacity) gets enqueued
and the thermals end right after the enqueue, the runqueue counter still
counts the task the small task as a giant task up until the point where
update_history on the task gets the chance to re-evaluate its giant status.

Change-Id: I405dbde6924224b7c1fc81bc38fa5138ca20e316
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>