usb: dwc3-msm: Execute dwc3_msm_host_ss_powerdown() when dp mode none
Currently, the dwc3_msm_host_ss_powerdown() uses an if the check for
dp_state which if not 2 lane or 4 lane should bail out of the function.
However, commit cedcba156689 ("usb: dwc3-msm: Communicate DP mode to USB3
PHY driver") introduced the enum which has the value of NONE assigned to 1.
This forces the condition to bail out even when the dp is in none mode.

Fix this by modifying the if to bail out when mode is 2 or 4 lanes &
bail out if DP_NONE is programmed.

Change-Id: Iad61e7bc59928532a7f016f3e1436ee31c963365
Signed-off-by: Udipto Goswami <udipto.goswami@oss.qualcomm.com>