usb: dwc3: msm: improve safety of dwc3_msm_vbus_notifier()
During probe there could be some spurious eud notification which can
cause crash while checking for gadget speed. This is because eud is
registered via extcon and its agnostic of the usb stack and can start
sending notification as any point.

Fix this by adding a check for gadget to avoid the crash.

Change-Id: Ic13ab8a920cd7c291b4aad7cf7af797f1f53a4ad
Signed-off-by: Udipto Goswami <udipto.goswami@oss.qualcomm.com>