snd: usb: usb_audio_qmi_svc: Fix sb_notifier check of ep_index
Currently the sb_notifier has a gating condition where it check if the
ep_indexes match with the sideband's ep index only then it allows to
send the disconnect_ind to the audio dsp. However, this fails because
the direction field was not taken into account in the condition
therefore when the direction changes the 8th bit gets set and mismatches
with the index, ultimately doesn't call the disconnect_ind.

Fix this by adding a mask of 0x7F which will ignore the direction and
will compared only the ep index.

Change-Id: Ia49753c709679f585a2b1b08ddb0e4c8888d7aff
Signed-off-by: Udipto Goswami <udipto.goswami@oss.qualcomm.com>