spi: spi-msm-geni: Validate protocol against compatible string in DT
Add a check in spi_verify_proto() to ensure the protocol read from
hardware matches the expected protocol based on the device tree
'compatible' string.

This helps catch misconfigurations where the DT node specifies an
incompatible device type. For example:
- If the compatible string is "qcom,qspi-geni", the protocol must be
GENI_SE_QSPI.
- If the compatible string is "qcom,spi-geni", the protocol must be
GENI_SE_SPI.

If a mismatch is detected, an error is logged and -ENXIO is returned.

Change-Id: I7bdc7b7fa80d3f9883530fa0760f2f399831438b
Signed-off-by: Sachin Gupta <quic_sachgupt@quicinc.com>