build: bazel: Add signing/verity keys to output
Signing keys and certificates are required to sign kernel modules
and dm-verity hash, but these files are not generated in the current
le build.

Add a Bazel rule to generate them and add to the output of the le
distributions. Also add the scripts and pre-generated key required
to use them.

Change-Id: I186596b27d6a1b91ed7910b10da439f0dc7ae6d9
Signed-off-by: Ankur Matariya <amatariy@qti.qualcomm.com>