include: linux: Remove qseecom.h from soc-repo
qseecom.h is needed by qseecom driver which is now present out-of-tree
in securemsm-kernel and should be deleted from soc-repo.

Along with this, include/soc/qcom/qseecomi.h is also internal to qseecom
driver which is not explicitly needed in soc-repo now.

Do required cleanup changes in other drivers too to adapt the header
file removal.

Change-Id: Id3d9877efba6f3c50ab400d8907be95df17b5d81
Signed-off-by: Kuldeep Singh <kuldsing@qti.qualcomm.com>