sched/walt: Refine sync wakeup fastpath for NORMAL tasks
On platforms where prime CPUs exhibit high leakage, we may observe
unintended high utilization of these cores. One such scenario occurs
when a task is boosted to a prime CPU, and after the boost expires, it
performs a sync wakeup of another task.

To mitigate this, avoid selecting a prime CPU for sync wakeups if
the newly awakened task can be accommodated by lower-performance
clusters.

Change-Id: I475b5c13de62406cf8aa1239fa3ad8583d50d14d
Signed-off-by: Atul Pant <quic_atulpant@quicinc.com>