zram: qpace: Align zsmalloc allocation flags with upstream implementation
The qpace path doesn’t need to acquire the local lock in the zcomp
stream, which means we can be more aggressive when allocating zs
pages. Since this is the final step before swapping out, and we have
a fallback mechanism to abandon if requests become excessive, this
approach appears safe.

Change-Id: Ieb8bfa7066384f695e3a46441a14d91e244d8980
Signed-off-by: Zhenhua Huang <quic_zhenhuah@quicinc.com>