usb: dwc3: defer probe for charger extcon synchronization
Consider a situation where the USB probe occurs before the PMIC
charger probe. If the USB extcon depends on the PMIC charger,
then the USB probe must return -EAGAIN or -EDEFER to defer the
device probing so it can be retried later.

The existing approach does not update the ret value for extcon
failures, which retains its value from the last successful fetch.
This causes issues during the DWC probe by returning a positive
value, incorrectly marking the probe as failed. As a result, the
DWC driver never probes again, and we don’t see the UDC getting
created.

Fix this by using ret to read the return value from the extcon
register and fail the driver probe if it is not 0.

Change-Id: I6baafad8f08b5a8d62da4f418dfd1eeb89c7585f
Signed-off-by: Akash Kumar <akakum@qti.qualcomm.com>