mmc: sdhci-msm: Change HS mode frequency to 37.5 with level shifter
Since we are using floor ops for all our targets to avoid overclocking
issue and now if we request for 37Mhz for HS mode SD card on targets
with level shifter that will lead to the frequency being set to lower
frequency.
Change HS mode frequency to 37.5 with level shifter.

Change-Id: I239ab543ccbeb20f2a098fbc41997f46ae9b7b59
Signed-off-by: Sarthak Garg <quic_sartgarg@quicinc.com>
Signed-off-by: Manish Pandey <manish.pandey@oss.qualcomm.com>