usb: dwc3: dwc3-msm-ops: Add kprobe to intercept host exit
Currently there is a race in usb controller when it tear down from host
stack to peripheral. It seems while transition xhci might not have
fully tear down and the set_mode would execute and controller will go to
device mode. Now when the run_stop gets set and controller tried to
perform read on registers it will throw smmu error with far 0x00.

Fix this by adding a delay in exit kprobe of dwc3_host_exit() allowing
controller to recover to a proper state.

Change-Id: I048a25c967c2dab20048ffb4d522a0a9ff4c32c4
Signed-off-by: Udipto Goswami <udipto.goswami@oss.qualcomm.com>