soc: qcom: gic_intr_routing: Handle defective cpu case
There are certain scenarios where logical CPU is not equal to the
physical cpu in HLOS. In this case, correct mpidr of the cpu that is
referred by logical cpu number is to be used.

Handle this case by reading cpuid_mpidr for the cpus that are available.
Map this value with the logical CPU number and use it directly.

Change-Id: I7537e3fdd46ecb1864d7acfa9fad29731b83aad0
Signed-off-by: Lakshmi Sowjanya D <laksd@qti.qualcomm.com>